// src/components/ProductCard.d.ts
interface ProductCardProps {
    id: string;
    name: string;
    description: string;
    price: number;
    image: string;
    category: string;
}
export declare const ProductCard: ({ name, description, price, image, category }: ProductCardProps) => import("react/jsx-runtime").JSX.Element;
export {};